pyproject.toml
[project]
name = "surface-change-monitor"
version = "0.1.0"
description = "Near-real-time impervious surface change detection for flood model updating"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopandas>=1.1.3",
    "numpy>=2.4.3",
    "pyproj>=3.7.2",
    "pystac-client>=0.9.0",
    "python-dotenv>=1.2.2",
    "rasterio>=1.5.0",
    "requests>=2.32.5",
    "rioxarray>=0.22.0",
    "shapely>=2.1.2",
    "xarray>=2026.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--strict-markers", "-ra"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "responses>=0.26.0",
]
